final changes before 1.0 API freeze
 - attr: buttonSize conflicts, so moving to buttonDimensions
 - attr: rating is float in Android by default, so our attr will be float too
 - common-ify the function to setControlColor for all compound buttons

+++ sushilib/src/main/java/com/zomato/sushilib/organisms/ratings/SushiRatingBar.kt
@@ -59,7 +59,7 @@ class SushiRatingBar @JvmOverloads constructor(
         val ta = context.obtainStyledAttributes(attrs, R.styleable.SushiRatingBar, defStyleAttr, defStyleRes)
         val tagSize = ta.getInt(R.styleable.SushiRatingBar_tagSize, TagSize.LARGE)
         val attrTagType = ta.getInt(R.styleable.SushiRatingBar_tagType, TagType.CAPSULE)
-        rating = ta.getInt(R.styleable.SushiRatingBar_rating, 0)
+        rating = ta.getFloat(R.styleable.SushiRatingBar_rating, 0f).toInt()
 
         internalTagType = if (attrTagType == TagType.CAPSULE || attrTagType == TagType.CAPSULE_OUTLINE) {
             TagType.CAPSULE
